to poc versjoner av hurtigvalg, en med hurtigvalg inline og en med tab-meny på toppen av dialogpanel
