+++ src/components/Hurtigvelger/HurtigvelgerExpandable.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import EkspanderbartpanelBase from 'nav-frontend-ekspanderbartpanel/lib/ekspanderbartpanel-base';
-import { Tekst, tekster } from './tekster';
-import HurtigvelgerElement from './HurtigvelgerElement';
+import { Tekst } from './tekster';
 import styled from 'styled-components';
 import theme from '../../styles/personOversiktTheme';
 import { Undertittel } from 'nav-frontend-typografi';
+import HurtigvalgListe from './HurtigvalgListe';
 
 interface Props {
     send: (tekst: Tekst) => void;
@@ -20,21 +20,14 @@ const Style = styled.div`
     filter: drop-shadow(0 0 1px rgba(0, 0, 0, 0.7));
 `;
 
-const MarginBottom = styled.div`
-    margin-bottom: 6rem;
-`;
-
-function Hurtigvelger(props: Props) {
+function HurtigvelgerExpandable(props: Props) {
     return (
         <Style>
             <EkspanderbartpanelBase heading={<Undertittel>SEND HURTIGSVAR</Undertittel>} ariaTittel={'hurtigsvar'}>
-                {tekster.map(tekst => (
-                    <HurtigvelgerElement key={tekst.tittel} tekst={tekst} sendTekst={() => props.send(tekst)} />
-                ))}
-                <MarginBottom />
+                <HurtigvalgListe send={props.send} />
             </EkspanderbartpanelBase>
         </Style>
     );
 }
 
-export default Hurtigvelger;
+export default HurtigvelgerExpandable;
